Exceptions and Exceptions handlers for REST

+++ app/src/main/java/hexlet/code/demo/controller/UserController.java
@@ -1,5 +1,6 @@
 package hexlet.code.demo.controller;
 
+import hexlet.code.demo.exception.ResourceNotFoundException;
 import hexlet.code.demo.model.User;
 import hexlet.code.demo.repository.UserRepository;
 import jakarta.validation.Valid;
@@ -7,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -19,13 +21,11 @@ import java.util.List;
 @RestController
 @RequestMapping("/api/users")
 public class UserController {
+
     @Autowired
     private UserRepository userRepository;
-    /* Лучше так указывать статус, ResponseEntity устаревший и нужен
-    когда заголовки нужно выставлять вручную или
-    реализовывать какую то специфичную функциональность*/
+
     @GetMapping
-    @ResponseStatus(HttpStatus.OK)
     public List<User> getAllUsers(@RequestParam(defaultValue = "10") Integer limit) {
         return userRepository.findAll().stream().limit(limit).toList();
     }
@@ -33,14 +33,21 @@ public class UserController {
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public User createUser(@Valid @RequestBody User user) {
-        var savedUser = userRepository.save(user);
-        return savedUser;
+        return userRepository.save(user);
     }
 
-    @DeleteMapping
+    @GetMapping("/{id}")
+    public User getUser(@PathVariable Long id) {
+        return userRepository.findById(id)
+                             .orElseThrow(() -> new ResourceNotFoundException("User with id " + id + " not found"));
+    }
+
+    @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteUser(@Valid @RequestParam Long id) {
+    public void deleteUser(@PathVariable Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new ResourceNotFoundException("User with id " + id + " not found");
+        }
         userRepository.deleteById(id);
     }
-
 }
